Assignment-2/requirements.txt
asgiref==3.8.1
Django==5.1.1
django-debug-toolbar==4.4.6
django-redis==5.4.0
redis==5.1.1
sqlparse==0.5.1
gunicorn
